#45  Move create index to DDL package
