+++ cql-ast/src/main/scala/troy/cql/ast/ddl/Index.scala
@@ -1,0 +1,11 @@
+package troy.cql.ast.ddl
+
+import troy.cql.ast.MapLiteral
+
+object Index {
+  case class Using(using: String, options: Option[MapLiteral])
+
+  trait IndexIdentifier
+  case class Identifier(value: String) extends IndexIdentifier
+  case class Keys(of: String) extends IndexIdentifier
+}
