+++ cql-parser/src/main/scala/troy/cql/parser/ddl/CreateIndexParser.scala
@@ -1,0 +1,31 @@
+package troy.cql.parser.ddl
+
+import troy.cql.ast.CqlParser._
+import troy.cql.ast.CreateIndex
+import troy.cql.ast.ddl.Index._
+
+trait CreateIndexParser {
+  def createIndex: Parser[CreateIndex] = {
+    def indexName = identifier.?
+    def onTable = "ON".i ~> tableName
+    def indexIdentifier: Parser[IndexIdentifier] = {
+      val keys = "KEYS".i ~> parenthesis(identifier) ^^ Keys
+      val ident = identifier ^^ Identifier
+      parenthesis(((keys | ident)))
+    }
+    def using = {
+      def withOptions =
+        "WITH".i ~> "OPTIONS".i ~> "=" ~> mapLiteral
+
+      "using".i ~> Constants.string ~ withOptions.? ^^^^ Using
+    }.?
+
+    "CREATE".i ~>
+      ("CUSTOM".flag <~ "INDEX".i) ~
+      ifNotExists ~
+      indexName ~
+      onTable ~
+      indexIdentifier ~
+      using ^^^^ CreateIndex.apply
+  }
+}
